Clean up tests for SE3 functional.
