+++ tests/labs/lie_functional/common.py
@@ -9,6 +9,67 @@ import torch
 TEST_EPS = 5e-7
 
 
+def get_test_cfg(op_name, dtype, dim, data_shape, module=None):
+    atol = TEST_EPS
+    # input_type --> tuple[str, param]
+    # input_types --> a tuple of type info for a given function
+    # all_input_types --> a list of input types, if more than one check isn needed
+    all_input_types = []
+    if op_name in ["exp", "hat"]:
+        all_input_types.append((("tangent", dim),))
+        atol = 1e-6 if op_name == "exp" else TEST_EPS
+    if op_name == "log":
+        all_input_types.append((("group", module),))
+        atol = 5e-6 if dtype == torch.float32 else TEST_EPS
+    if op_name in ["adjoint", "inverse"]:
+        all_input_types.append((("group", module),))
+    if op_name == "compose":
+        all_input_types.append((("group", module),) * 2)
+    if op_name == "lift":
+        matrix_shape = (torch.randint(1, 20, ()).item(), dim)
+        all_input_types.append((("matrix", matrix_shape),))
+    if op_name == "project":
+        matrix_shape = (torch.randint(1, 20, ()).item(),) + data_shape
+        all_input_types.append((("matrix", matrix_shape),))
+    if op_name == "left_act":
+        for shape in [
+            (3, torch.randint(1, 5, ()).item()),
+            (2, 4, 3, torch.randint(1, 5, ()).item()),
+        ]:
+            all_input_types.append((("group", module), ("matrix", shape)))
+    if op_name == "left_project":
+        for shape in [
+            data_shape,
+            ((torch.randint(1, 5, ()).item(),) + data_shape),
+        ]:
+            all_input_types.append((("group", module), ("matrix", shape)))
+    return all_input_types, atol
+
+
+# Sample inputs with the desired types.
+def sample_inputs(input_types, batch_size, dtype, rng, module=None):
+    def _sample(input_type):
+        type_str, param = input_type
+
+        def _quat_sample():
+            q = torch.rand(batch_size, param, dtype=dtype, generator=rng)
+            return q / torch.norm(q, dim=1, keepdim=True)
+
+        sample_fns = {
+            "tangent": lambda: torch.rand(
+                batch_size, param, dtype=dtype, generator=rng
+            ),
+            "group": lambda: param.rand(batch_size, generator=rng, dtype=dtype),
+            "quat": lambda: _quat_sample(),
+            "matrix": lambda: torch.rand(
+                (batch_size,) + param, generator=rng, dtype=dtype
+            ),
+        }
+        return sample_fns[type_str]()
+
+    return tuple(_sample(type_str) for type_str in input_types)
+
+
 # Checks if the jacobian computed by default torch autograd is close to the one
 # provided with custom backward
 # funcs is a list of callable that modifiies the jacobian. If provided we also
