+++ tests/labs/lie_functional/test_so3.py
@@ -8,71 +8,13 @@ import pytest
 import torch
 
 from tests.decorators import run_if_labs
-from .common import TEST_EPS, check_lie_group_function, left_project_func
-
-
-def _get_inputs(input_types, batch_size, dtype, rng, module=None):
-    def _sample(input_type):
-        type_str, param = input_type
-
-        def _quat_sample():
-            q = torch.rand(batch_size, param, dtype=dtype, generator=rng)
-            return q / torch.norm(q, dim=1, keepdim=True)
-
-        sample_fns = {
-            "tangent": lambda: torch.rand(
-                batch_size, param, dtype=dtype, generator=rng
-            ),
-            "group": lambda: param.rand(batch_size, generator=rng, dtype=dtype),
-            "quat": lambda: _quat_sample(),
-            "matrix": lambda: torch.rand(
-                (batch_size,) + param, generator=rng, dtype=dtype
-            ),
-        }
-        return sample_fns[type_str]()
-
-    return tuple(_sample(type_str) for type_str in input_types)
-
-
-def _get_test_cfg(op_name, dtype, module=None):
-    atol = TEST_EPS
-    # input_type --> tuple[str, param]
-    # input_types --> a tuple of type info for a given function
-    # all_input_types --> a list of input types, if more than one check isn needed
-    all_input_types = []
-    if op_name == "exp":
-        all_input_types.append((("tangent", 3),))
-        atol = 1e-6
-    if op_name == "log":
-        all_input_types.append((("group", module),))
-        atol = 5e-6 if dtype == torch.float32 else TEST_EPS
-    if op_name in ["adjoint", "inverse"]:
-        all_input_types.append((("group", module),))
-    if op_name in ["hat"]:
-        all_input_types.append((("tangent", 3),))
-    if op_name == "compose":
-        all_input_types.append((("group", module),) * 2)
-    if op_name == "quaternion_to_rotation":
-        all_input_types.append((("quat", 4),))
-    if op_name == "lift":
-        matrix_shape = (torch.randint(1, 20, ()).item(), 3)
-        all_input_types.append((("matrix", matrix_shape),))
-    if op_name == "project":
-        matrix_shape = (torch.randint(1, 20, ()).item(), 3, 3)
-        all_input_types.append((("matrix", matrix_shape),))
-    if op_name == "left_act":
-        for shape in [
-            (3, torch.randint(1, 5, ()).item()),
-            (2, 4, 3, torch.randint(1, 5, ()).item()),
-        ]:
-            all_input_types.append((("group", module), ("matrix", shape)))
-    if op_name == "left_project":
-        for shape in [
-            (3, 3),
-            (torch.randint(1, 5, ()).item(), 3, 3),
-        ]:
-            all_input_types.append((("group", module), ("matrix", shape)))
-    return all_input_types, atol
+from .common import (
+    TEST_EPS,
+    check_lie_group_function,
+    get_test_cfg,
+    left_project_func,
+    sample_inputs,
+)
 
 
 @run_if_labs()
@@ -100,9 +42,9 @@ def test_op(op_name, batch_size, dtype):
     rng = torch.Generator()
     rng.manual_seed(0)
 
-    all_input_types, atol = _get_test_cfg(op_name, dtype, module=so3)
+    all_input_types, atol = get_test_cfg(op_name, dtype, 3, (3, 3), module=so3)
     for input_types in all_input_types:
-        inputs = _get_inputs(input_types, batch_size, dtype, rng)
+        inputs = sample_inputs(input_types, batch_size, dtype, rng)
         funcs = (
             tuple(left_project_func(so3, x) for x in inputs)
             if op_name == "log"
